requirements.txt
flask
flask-login
flask-bcrypt
flask-mail
flask-wtf
flask-sqlalchemy
sqlalchemy-migrate
argparse
